Lots of work on toal tools and elemental attack system

+++ src/main/java/com/github/gresh113/bionimine/capabilities/IToaEnergy.java
@@ -2,9 +2,9 @@ package com.github.gresh113.bionimine.capabilities;
 
 public interface IToaEnergy {
 	int getKanohiEnergy();
-	byte getElementalEnergy();
+	int getElementalEnergy();
 	
 	void setKanohiEnergy(int energy);
-	void setElementalEnergy(byte energy);
+	void setElementalEnergy(int energy);
 	
 }
